Updated all needed functions
Now, can commute between manual and automatic mode.
In automatic mode, upon scanning RFID tag, the master checks the current season and the current temperature in the room, compares it to the preferred temperature of the user, and makes a decision.
In manual mode, the master sends a message on topic_mode topic to let the board responsible for the manual temperature know the mode switched to manual and the manual temperature is required. After sending the manual temperature to the master board, the master also makes a decision similar to the one in the automatic mode, and depending on the current season and current temperature, sends a command to the corresponding board.